Change default toast duration to ToastDuration.Long
Updated the ShowToastAsync method in the ToastService class to change the default value of the duration parameter from ToastDuration.Short to ToastDuration.Long. This ensures that toasts will display for a longer duration by default if the duration parameter is not specified.

+++ ec.com.naturisa.mobile.feedcontrol/Services/Notifications/ToastService .cs
@@ -4,7 +4,7 @@
     {
         public async Task ShowToastAsync(
             string message,
-            ToastDuration duration = ToastDuration.Short,
+            ToastDuration duration = ToastDuration.Long,
             double fontSize = 14
         )
         {
